+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/Generators.scala
@@ -3,12 +3,13 @@ package org.broadinstitute.dsde.workbench.google2
 import java.nio.charset.Charset
 import java.time.Instant
 
+import com.google.pubsub.v1.ProjectTopicName
 import org.broadinstitute.dsde.workbench.model.google._
 import org.scalacheck.Gen
 
 object Generators {
   val utf8Charset = Charset.forName("UTF-8")
-  val genGoogleProject: Gen[GoogleProject] = Gen.uuid.map(x => GoogleProject(s"bucket${x.toString}"))
+  val genGoogleProject: Gen[GoogleProject] = Gen.uuid.map(x => GoogleProject(s"project${x.toString}"))
   val genGcsBucketName: Gen[GcsBucketName] = Gen.uuid.map(x => GcsBucketName(s"bucket${x.toString}"))
   val genGcsBlobName: Gen[GcsBlobName] = Gen.uuid.map(x => GcsBlobName(s"object${x.toString}"))
   val genGcsObjectName: Gen[GcsObjectName] = for{
@@ -16,4 +17,13 @@ object Generators {
     createdTime <- Gen.calendar.map(x => Instant.ofEpochMilli(x.getTimeInMillis))
   } yield GcsObjectName(blobName.value, createdTime)
   val genGcsObjectBody: Gen[Array[Byte]] = Gen.alphaStr.map(x => x.getBytes(utf8Charset))
+  val genPerson: Gen[Person] = for{
+    name <- Gen.alphaStr.map(s => s"name$s")
+    email <- Gen.alphaStr.map(s => s"email$s")
+  } yield Person(name, email)
+  val genListPerson = Gen.nonEmptyListOf(genPerson)
+  val genProjectTopicName: Gen[ProjectTopicName] = for{
+    project <- genGoogleProject
+    topic <- Gen.alphaStr.map(x => s"topic$x")
+  } yield ProjectTopicName.of(project.value, topic)
 }