+++ util/src/test/scala/org/broadinstitute/dsde/workbench/util/TestUtil.scala
@@ -1,10 +1,14 @@
 package org.broadinstitute.dsde.workbench.util
 
-import cats.effect.IO
+import cats.effect.{ContextShift, IO, Timer}
 import org.scalatest.Assertion
 
 import scala.concurrent.Future
+import scala.concurrent.ExecutionContext.global
 
-object TestUtil {
+trait WorkbenchTest {
+  implicit val timer: Timer[IO] = IO.timer(global)
+  implicit val cs: ContextShift[IO] = IO.contextShift(global)
+
   def ioAssertion(test: => IO[Assertion]): Future[Assertion] = test.unsafeToFuture()
 }